Move all applications api to applications folder
