+++ components/ApplicationSteps/Step1.tsx
@@ -2,7 +2,7 @@ import { Dispatch, SetStateAction, useEffect, useState } from 'react'
 import { ApplicationData } from '../../classes/application_data'
 import { faculties } from '../../constants/faculties'
 import { useAuth } from '../../context/AuthUserContext'
-import { updateApplicationData } from '../../pages/api/step1'
+import { updateApplicationData } from '../../pages/api/applications/step1'
 import FileUploadComponent from './FileUpload'
 import ProceedButtons from './ProceedButtons'
 import SelectInput from './SelectInput'
