+++ components/ApplicationSteps/Step2.tsx
@@ -2,8 +2,8 @@ import { Dispatch, SetStateAction, useEffect, useState } from 'react'
 import { ApplicationData } from '../../classes/application_data'
 import { faculties } from '../../constants/faculties'
 import { useAuth } from '../../context/AuthUserContext'
-import { updateApplicationData } from '../../pages/api/step2'
-import { deleteDocuments } from '../../pages/api/uploadDocument'
+import { updateApplicationData } from '../../pages/api/applications/step2'
+import { deleteDocuments } from '../../pages/api/applications/uploadDocument'
 import { AcademicRecordType } from '../../types'
 import CheckBoxInput from './Checkboxes'
 import FileUploadComponent from './FileUpload'
