+++ components/ApplicationSteps/Step4.tsx
@@ -1,8 +1,8 @@
 import { Dispatch, SetStateAction, useEffect, useState } from 'react'
 import { ApplicationData } from '../../classes/application_data'
 import { useAuth } from '../../context/AuthUserContext'
-import { updateApplicationData } from '../../pages/api/step4'
-import { deleteDocuments } from '../../pages/api/uploadDocument'
+import { updateApplicationData } from '../../pages/api/applications/step4'
+import { deleteDocuments } from '../../pages/api/applications/uploadDocument'
 import { WorkExperiencesType } from '../../types'
 import CheckBoxInput from './Checkboxes'
 import FileUploadComponent from './FileUpload'
