+++ components/ApplicationSteps/Step5.tsx
@@ -1,8 +1,8 @@
 import { Dispatch, SetStateAction, useEffect, useState } from 'react'
 import { ApplicationData } from '../../classes/application_data'
 import { useAuth } from '../../context/AuthUserContext'
-import { updateApplicationData } from '../../pages/api/step5'
-import { deleteDocuments } from '../../pages/api/uploadDocument'
+import { updateApplicationData } from '../../pages/api/applications/step5'
+import { deleteDocuments } from '../../pages/api/applications/uploadDocument'
 import { PosterOrWorkshopsType } from '../../types'
 import FileUploadComponent from './FileUpload'
 import ProceedButtons from './ProceedButtons'
