+++ components/ApplicationSteps/Step7.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, SetStateAction, useEffect, useState } from 'react'
 import { ApplicationData } from '../../classes/application_data'
 import { useAuth } from '../../context/AuthUserContext'
-import { updateApplicationData } from '../../pages/api/step7'
+import { updateApplicationData } from '../../pages/api/applications/step7'
 import { ExtraCurricularsType } from '../../types'
 import FileUploadComponent from './FileUpload'
 import ProceedButtons from './ProceedButtons'
