+++ pages/api/applications/step1.ts
@@ -1,5 +1,5 @@
 import path from 'path'
-import { firestore } from '../../firebase'
+import { firestore } from '../../../firebase'
 
 export const updateApplicationData = (
   userId: string,
